backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
langchain
langchain-openai
langchain-community
pydantic
python-multipart
chromadb
